yes
Eric Witowski
4713 W Shore Dr
Mchenry, Illinois 60050

To Whom It May Concern:


 	I wanted to make my cover letter a glimpse of the person you would consider for this position.  As a resume is created for experience and education, the cover letter is about looking at characteristics of who or what kind of person or applicant you would be hiring.  I say this because a person’s ability to think outside the box cannot be placed on paper, a person’s personality or understanding of inner office workflow cannot be gauged by a resume.  The real question is will I the candidate this excel at completing any and all task assigned.  To this question I say yes that I do have the ability to establish great relationships with the internal customer and external customers while also completing assigned tasks given within the scope of any duties assigned within any company or organization in a timely fashion.  

My current career objective is to become extremely knowledgeable in the new work environment/infrastructure of the company or organization being hired into.  To apply my knowledge and experience to my new position being applied for and to quickly adapt and learn OJT skills in the workplace environment.  I am an experienced line manager with extensive knowledge in team-building that ranges from organizing project teams while also encouraging individuals in the team to work together in current or future challenges with motivation and inspiration and teamwork.  I use an Agile approach to my daily tasks, and some would say that this is like writing down goals on a post-it notes but in a more advanced form.  The advantage of using the Agile methodology allows me to keep tabs on each person in the team and the group as a whole using techniques like standups for the daily morning meetings.   I have developed many solid skills in hiring and retaining employees while also having incorporated technological advances into a company where such advances require a significant amount of employee training, and knowledge upgrading.   

     In conclusion, I believe I have the required skills, knowledge, and experience for this position. I am a person that works great in teams, as their leader or an individual.  I can apply as much or as little experience and effort into any position hired into, as well as the technological savvy and a positive attitude toward implementing/incorporating change.  I am skilled in the technological and administrative aspects of the business enterprise and with these skills I have an added advantage to be the best candidate for this position.  


Thank You and Kindest Regards,
Eric Witowski
847-262-0311